move files, remove filters
